navbar complete
