+++ src/CommonClasses/base/base.scss
@@ -1,5 +1,5 @@
 // theme scss file
-@import url(../theme/theme.scss);
+@import "../theme/theme.scss";
 // typo scss file
 @import url(../typo/typo.scss);
 // button sccc file
