+++ src/CommonClasses/button/button.scss
@@ -1,5 +1,5 @@
 @import "../theme/theme.scss";
-// $primary-color: red;
+
 // main class of  button
 .button {
   width: fit-content;
