+++ src/CommonClasses/typo/typo.scss
@@ -1,3 +1,8 @@
+// font family links
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;400;500;600;700;800;900&family=Syne:wght@500;600&display=swap");
+.poppins {
+  font-family: "Poppins", sans-serif;
+}
 // fomt size scss
 
 .text-3xs {
