+++ src/components/navbar.scss
@@ -1,151 +1,132 @@
-@import url("https://fonts.googleapis.com/css2?family=Syne:wght@500;600&display=swap");
+// @import url("https://fonts.googleapis.com/css2?family=Syne:wght@500;600&display=swap");\
+@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;400;500;600;700;800;900&family=Syne:wght@500;600&display=swap");
 @import url(../CommonClasses/base/base.scss);
+
 * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-  font-family: "Syne", sans-serif;
 }
+
+/* navbar style start  */
+
 .main-nav {
   width: 100%;
-  display: grid;
-  grid-template-columns: 96px 1fr 1fr 96px;
   height: 87px;
-  position: fixed;
+  display: grid;
+  align-items: center;
+  grid-template-columns: 96px 1fr 1fr 96px;
   background: linear-gradient(
     180deg,
     #ffffff 77.69%,
     rgba(255, 255, 255, 0.19) 101.79%
   );
+  position: fixed;
+  // box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,
+  //   rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;
   @media (max-width: 1200px) {
     grid-template-columns: 40px 1fr 1fr 40px;
   }
   @media (max-width: 992px) {
-    grid-template-columns: 40px 1fr 2fr 40px;
+    grid-template-columns: 20px 1fr 2fr 20px;
   }
   @media (max-width: 768px) {
-    grid-template-columns: 20px 1fr auto 10px;
+    grid-template-columns: 1fr auto;
+    background: #a9dadb;
   }
-  .ham-div {
-    align-items: center;
+}
+.hamburger-menu {
+  align-items: center;
+  display: none;
+  height: 87px;
+  justify-content: end;
+  @media (max-width: 768px) {
+    display: flex;
+    padding-right: 10px;
+  }
+}
+
+.logo {
+  display: grid;
+  justify-content: start;
+  align-items: center;
+  grid-column: 2/3;
+  height: 87px;
+  @media (max-width: 768px) {
+    grid-column: 1/2;
+    padding-left: 10px;
+    padding-top: 15px;
+  }
+}
+
+.menu-link {
+  grid-column: 3/4;
+  @media (max-width: 768px) {
     display: none;
-    height: 87px;
-    justify-content: end;
-    @media (max-width: 768px) {
-      display: flex;
-    }
   }
-  .logo-div {
+}
+.mobile-menu-link {
+  @media (max-width: 768px) {
+    grid-column: 2/4;
+    position: relative;
+    z-index: 99;
+    background-color: white;
+    height: 30rem;
     display: grid;
-
-    justify-content: start;
+    grid-column: 1/4;
     align-items: center;
-    grid-column: 2/3;
-    @media (max-width: 768px) {
-      grid-column: 1/2;
-      padding-left: 10px;
-    }
-  }
-
-  .nav-link {
-    @media (max-width: 768px) {
-      grid-column: -5/-1;
-      background: white;
-    }
+    padding-left: 3rem;
+    transition: all 2s linear;
+    transform-origin: top;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,
+      rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;
     .nav-link-wrapper {
-      // display: flex;
-      justify-content: space-between;
+      height: 28rem;
+      display: flex;
+      justify-content: space-around;
+      flex-direction: column;
       align-items: center;
-      height: 87px;
-      @media (max-width: 768px) {
-        flex-direction: column;
-        gap: 10px;
-        height: 70vh;
-      }
-      .nav-item {
-        list-style: none;
-        .nav-item-div {
-          text-decoration: none;
-          color: black;
-        }
+      .nav-item &:first-child {
+        transition-delay: 0.2s;
       }
     }
   }
 }
 
-// .project-navbar {
-//   width: 100%;
-//   height: 87px;
-//   background: linear-gradient(
-//     180deg,
-//     #ffffff 77.69%,
-//     rgba(255, 255, 255, 0.19) 101.79%
-//   );
-//   margin-left: 77px;
-//   padding-top: 16px;
-//   padding-bottom: 9px;
-//   position: fixed;
-//   @media (max-width: 600px) {
-//     padding-top: 15px;
-//     background: #00aaff75;
-//   }
-
-//   #checkbtn {
-//     float: right;
-//     line-height: 70px;
-//     display: none;
-//     padding-right: 20px;
-//     @media (max-width: 768px) {
-//       display: block;
-//     }
-//   }
-//   #check {
-//     display: none;
-//   }
-//   #check :checked ~ .project-function {
-//     left: 0;
-//   }
-//   .logo-div {
-//     margin-left: 10px;
-//     @media (max-width: 1200px) {
-//       margin-left: 40px;
-//     }
-//     @media (max-width: 992px) {
-//       margin-left: 20px;
-//     }
-//   }
-//   .project-function {
-//     float: right;
-//     list-style: none;
-//     display: flex;
-//     justify-content: space-between;
-//     width: 55%;
-//     line-height: 60px;
-
-//     @media (max-width: 1200px) {
-//       margin-right: 40px;
-//     }
-//     @media (max-width: 992px) {
-//       margin-right: 20px;
-//     }
-//     @media (max-width: 768px) {
-//       width: 100%;
-//       height: 100%;
-//       margin-right: 0;
-//       padding: 20px;
-//       flex-direction: column;
-//       align-items: center;
-//       position: fixed;
-//       top: 87px;
-//       left: -100%;
-//     }
-//     .home-button {
-//       font-weight: 500;
-
-//       background: linear-gradient(90deg, #b295f1 -32.14%, #0184ff 140.48%);
-//       -webkit-background-clip: text;
-//       -webkit-text-fill-color: transparent;
-//       background-clip: text;
-//     }
-//   }
-// }
+.nav-link-wrapper {
+  // height: 10rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  @media (max-width: 1200px) {
+    height: 8rem;
+  }
+  .nav-item {
+    list-style: none;
+    .nav-item-div {
+      font-family: "Poppins", sans-serif;
+      text-decoration: none;
+      font-size: 14px;
+      font-weight: 600;
+      text-transform: capitalize;
+      color: rgba(0, 0, 3, 0.8);
+      transition: 0.5s;
+      &:hover {
+        transform-origin: left;
+        color: rgba(0, 0, 3, 1);
+        text-decoration: underline;
+        transition: 0.5s;
+      }
+    }
+    .home-active {
+      background: linear-gradient(
+        90deg,
+        #b295f1 -32.14%,
+        rgba(1, 132, 255, 0.4) 140.48%
+      );
+      -webkit-background-clip: text;
+      -webkit-text-fill-color: transparent;
+      background-clip: text;
+    }
+  }
+}
